Refactor source code + fix some issues
* Move Goml code from goml/goml so that it can be imported as
`github.com/JulzDiverse/goml`
* Split `goml.go` into separate files for each ones of the operations
* Split tests into different files for each one of the operations
* Operations over files are exported as part of the library `GetFromFile`, `TransferToFile`, `DeleteInFile` and `SetInFile`
* Transfer can be reused as part of the library, `Transfer` and
`TransferToFile`
* DeleteInFile now removes the file if yaml is an empty file
* Fix wrong tests: most of them were not asserting the right thing
