+++ .gitignore
@@ -1,3 +1,4 @@
 goml-linux-amd64
 goml-darwin-amd64
 vendor/*
+goml
