+++ Makefile
@@ -1,0 +1,4 @@
+all:
+	go build -o goml cmd/goml/main.go
+test:
+	go test
